Client Update (Modular Test Client)
When I first I created a Test client that was very quick and not very
easy to test or modify just to see if I could get my ProblemModule
concept to work but having a flawed client could really cause some
trouble; so I added a new class Called ModularTestClient that you can
use completely from JUnit.

I'll do the same with the Server and Nodes, so when we start
implementing features we can quickly and easily test that we didn't
break anything.