BeforeBed Commit
Made some changes to a few areas. Mainly looked into the best way to
manage connections. This is much more involved multithreading than we
have done in the past, so the server has to be threadsafe all the way
through.

I am going to use an ExecutorService on a cached thread pool to handle
resource allocation and deallocation, as well as thread creation and
thread reclaimation.

A LinkedBlockingQueue to handle Tasks, and a Decorator pattern to accept
all incoming connections as a base connection then decorate them as
either a Node or a Client based on what they send. If the first thing
that they send is a ProblemModule, they are a client. If the first thing
they send is the Status Packet they are a node. Wrap the connection in
the proper decorator and put it into its own blocking Hashtable(may
change structure), since you don't want to have two threads send jobs to
the same nodes at the same time would crash the Node socket so you
bottleneck it at filter which is a short process which should maximize
the grid usage.