Before bed commit
It took a long time but I figured it out, In order to keep all of the
requisite connections alive and still being able to send/receive the
ProblemModules as soon as they are ready, I had to multi thread the
Server AND the nodes, the theoretical problem was a Node getting a
problem it would take a long time to compute the result for (i.e
rendering). I setup the node with 3 threads, one thread looks for
incoming Problems, one thread Solves the pending problem, and the main
process sends status packets to 1 keep the connection alive and 2 let
the server know if its ready for a new problem.

I also created a server / node test pair that works currently. I also
made more progress on the MultiThreadedServerConnectionManager. I know
how I'm gonna get the server to handle all of the connections and handle
multiple problems. I'm planning on making multiple problem management
threads to manage the pipeline of a problem from the queue to scheduling
the nodes and the reassembly and return to the client.

So it will be one thread per connection and 1-4 threads to handle queued
problems, so lots going on all at the same time.

If you guys have questions let me know. I've been just kind of working
to get the base cases working so I know how to best expand on it.