Light documentation
I added some comments and TODO: tags. Currently we have 48 TODO tags,
and 5 warnings in the whole project.

About half of the TODO tags are small  tasks so we should be able to
knock them out quick.

4 of the 5 warnings are variables not used locally and they have TODO
tags to use them. The last is we need to parameterize a callable but I
didn't know what It needed to return yet.